use resized campfire image, add meta tags for search engines, social media, and general embeds, add title, add viewport meta
The viewport meta should increase performance/percieved load time for users. Title is required meta, the rest is nice stuff that will boost SEO and allow nice embeds. The appropriately sized image should offer a noticeable performance increase, but should eventually be swapped for an appropriate format such as webp for *ultimate* performance. I also modified the header to use the smaller image though I am not sure if it is utilized.